refactor(intents): refactor business logic to service layer
